requirements.txt
google-generativeai==0.8.3
playwright==1.49.0
Pillow==10.4.0
numpy==1.26.4
python-dotenv==1.0.1
cairosvg==2.7.1
requests==2.32.3
